config: change folder structure for frontend
