+++ 3d-virtual-nlu/src/App.tsx
@@ -1,5 +1,5 @@
 import "./App.css";
-import RouterConfig from "./components/Router";
+import RouterConfig from "./Router";
 import { BrowserRouter as Router } from "react-router-dom";
 
 function App() {
