--- 3d-virtual-nlu/src/components/Admin/Navigation/navigation.module.css
@@ -1,19 +1,0 @@
-.navigation {
-  top: 0;
-  width: 20%;
-  min-height: 100vh;
-  color: #fff;
-  display: flex;
-  flex-direction: column;
-  background-color: rgb(83, 181, 44);
-}
-
-.avatar {
-}
-
-.profileImage {
-  padding: 0;
-  object-fit: cover;
-  border-radius: 50%;
-  border: 2px solid #000;
-}
