--- 3d-virtual-nlu/src/components/Admin/Tours/Tours.tsx
@@ -1,7 +1,0 @@
-import React from "react";
-
-const Tours = () => {
-  return <div>Tours</div>;
-};
-
-export default Tours;
