--- 3d-virtual-nlu/src/components/HomePage/CampusMap/CampusMap.tsx
@@ -1,12 +1,0 @@
-import React from "react";
-import styles from "./CampusMap.module.css";
-
-const CampusMap = () => {
-  return (
-    <section id="campusMap" className={styles.campusMapContainer}>
-      <h2 className={styles.title}>Sơ đồ trường</h2>
-    </section>
-  );
-};
-
-export default CampusMap;
