--- 3d-virtual-nlu/src/components/TourManager/CreateTour/CreateTour.tsx
@@ -1,7 +1,0 @@
-import React from "react";
-
-const CreateTour = () => {
-  return <div>CreateTour</div>;
-};
-
-export default CreateTour;
