--- 3d-virtual-nlu/src/components/TourManager/EditTour/EditTour.tsx
@@ -1,7 +1,0 @@
-import React from "react";
-
-const EditTour = () => {
-  return <div>EditTour</div>;
-};
-
-export default EditTour;
