+++ 3d-virtual-nlu/src/components/visitor/CampusMap.tsx
@@ -1,0 +1,12 @@
+import React from "react";
+import styles from "../../styles/campusMap.module.css";
+
+const CampusMap = () => {
+  return (
+    <section id="campusMap" className={styles.campusMapContainer}>
+      <h2 className={styles.title}>Sơ đồ trường</h2>
+    </section>
+  );
+};
+
+export default CampusMap;
