+++ 3d-virtual-nlu/src/main.tsx
@@ -3,7 +3,7 @@ import { createRoot } from "react-dom/client";
 import "./index.css";
 import App from "./App.tsx";
 
-import { store } from "./components/redux/store";
+import { store } from "./redux/store.tsx";
 import { Provider } from "react-redux";
 
 createRoot(document.getElementById("root")!).render(
